Create <LinearStream> components that visually streams like PHP
In practice it is better than PHP because sibling suspense boundaries can still execute.

I think there is an argument that Next should wrap layout/layout/page boundaries with a component like this. If we don't do that, we should still make a component library that provides reusable non-primitive streaming functionality.