Use an optimistic cart count from localStorage
While we do have cookies in this case, the real cart count is more likely
stored in a database. To avoid the flash of a missing cart count we can
stash the previous cart count in localStorage and optimistically use that
value from an inline script to ensure that we instantly show a cart count
which will be correct assuming the user hasn't added to the cart else
where.

This can also be done with things like logged in states like username/profile.