Migrate to `dynamicIO`
For the `useLinkStatus` example I had to add a loading boundary, which
is required by Dynamic IO for a fully dynamic page. This means that the
inline loading indicator is only shown when starting on the home page,
and then navigating to the `useLinkStatus` page and its child pages.
When starting at `/use-link-status`, the loading boundary is available
to the client router, and will be used instead of the inline indicator
when switching between child pages.